# Improve `grace_hash` join by reserving hash table's size