<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/cancel_and_rerun_workflow_lambda/app.py
-->
### Changelog category (leave one):
- Performance Improvement



### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

Try to reserve hash table's size in `grace_hash` join.

for following query
```SQL
CREATE TABLE t1 (key UInt32, s String) engine = TinyLog;
CREATE TABLE t2 (key UInt32, s String) engine = TinyLog;


insert into t1 select number as key, toString(number) as s from numbers(20000000);
insert into t2 select number as key, toString(number/3) as s from numbers(20000000);
select count(1) from (SELECT key FROM t1 AS t1 ALL LEFT JOIN t2 AS t2 USING (key) )t1 settings join_algorithm='grace_hash', max_rows_in_join=1000000;
```
before
```
1 row in set. Elapsed: 20.685 sec. Processed 40.00 million rows, 160.00 MB (1.93 million rows/s., 7.74 MB/s.)
```
after
```
1 row in set. Elapsed: 16.146 sec. Processed 40.00 million rows, 160.00 MB (2.48 million rows/s., 9.91 MB/s.)
```

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->


> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/