@vdimir 2023-05-24
Great!
<img width="1211" alt="image" src="https://github.com/ClickHouse/ClickHouse/assets/7023786/b5613608-9dc0-475a-b991-eb602464d75a">