@vdimir 2023-05-24
> Is there a chance it could break join_set_filter? https://s3.amazonaws.com/clickhouse-test-reports/0/3f892ceb12b868b1b9deb0607a7df2dde5f1a139/performance_comparison_[1_4]/report.html

@tavplubix It shouldn't, I'll check the reason